==> ipynb/2.1 Preprocessing sur le csv application.ipynb <==
# J'installe les librairies dont je vais avoir besoin. 

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt  
import seaborn as sns

# On importe le csv application_data
application = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_Originaux/application_data.csv')
                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         

# On indique vouloir garder toutes les colonnes
pd.set_option('display.max_columns',None)

# On trouve 122 colonnes 
application.shape

# On calcule la corrélation par paire des colonnes, à l'exclusion des valeurs NA/null.
coorMatrix = application.corr(method='pearson')
coorMatrix.shape

# On à un problème on passe de 122 colonnes à 106 ...en effet, il y a des string qui ne nous permettent pas
# d'appliquer une matrice de corrélation : nous allons donc recoder toutes les variables string en int. 

# On vérifie toutes les données uniques dans la colonne GENDER
print(application['CODE_GENDER'].unique())

# On cherche a remplacer les variables où il y a des string par des variables numériques

application.NAME_CONTRACT_TYPE=application.NAME_CONTRACT_TYPE.replace(["Cash loans","Revolving loans"],[1,2])
application.CODE_GENDER=application.CODE_GENDER.replace(["M","F","XNA"],[1,2,3])
application.FLAG_OWN_CAR=application.FLAG_OWN_CAR.replace(["Y","N"],[1,2])
application.FLAG_OWN_REALTY=application.FLAG_OWN_REALTY.replace(["Y","N"],[1,2])
application.NAME_TYPE_SUITE=application.NAME_TYPE_SUITE.replace(['Unaccompanied','Family','Spouse, partner','Children','Other_A','Other_B','Group of people'],[1,2,3,4,5,6,7])
application.NAME_INCOME_TYPE=application.NAME_INCOME_TYPE.replace(['Working','State servant','Commercial associate','Pensioner','Unemployed','Student','Businessman','Maternity leave'],[1,2,3,4,5,6,7,8])
application.NAME_EDUCATION_TYPE=application.NAME_EDUCATION_TYPE.replace(['Secondary / secondary special','Higher education','Incomplete higher','Lower secondary','Academic degree'],[1,2,3,4,5])
application.NAME_FAMILY_STATUS=application.NAME_FAMILY_STATUS.replace(['Single / not married','Married','Civil marriage','Widow','Separated','Unknown'],[1,2,3,4,5,6])
application.NAME_HOUSING_TYPE=application.NAME_HOUSING_TYPE.replace(['House / apartment','Rented apartment','With parents','Municipal apartment','Office apartment','Co-op apartment'],[1,2,3,4,5,6])
application.OCCUPATION_TYPE=application.OCCUPATION_TYPE.replace(['Laborers','Core staff','Accountants','Managers','Drivers','Sales staff','Cleaning staff','Cooking staff','Private service staff','Medicine staff','Security staff','High skill tech staff','Waiters/barmen staff','Low-skill Laborers','Realty agents','Secretaries','IT staff','HR staff'],[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18])
application.WEEKDAY_APPR_PROCESS_START=application.WEEKDAY_APPR_PROCESS_START.replace(['WEDNESDAY','MONDAY','THURSDAY','SUNDAY','SATURDAY','FRIDAY','TUESDAY'],[3,4,1,7,6,5,2])
application.ORGANIZATION_TYPE=application.ORGANIZATION_TYPE.replace(['Business Entity Type 3','School','Government','Religion','Other','XNA','Electricity','Medicine','Business Entity Type 2','Self-employed','Transport: type 2','Construction','Housing','Kindergarten','Trade: type 7','Industry: type 11','Military','Services','Security Ministries','Transport: type 4','Industry: type 1','Emergency','Security','Trade: type 2','University','Transport: type 3','Police','Business Entity Type 1','Postal','Industry: type 4','Agriculture','Restaurant','Culture','Hotel','Industry: type 7','Trade: type 3','Industry: type 3','Bank','Industry: type 9','Insurance','Trade: type 6','Industry: type 2','Transport: type 1','Industry: type 12','Mobile','Trade: type 1','Industry: type 5','Industry: type 10','Legal Services','Advertising','Trade: type 5','Cleaning','Industry: type 13','Trade: type 4','Telecom','Industry: type 8','Realtor','Industry: type 6'],[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58])
application.FONDKAPREMONT_MODE=application.FONDKAPREMONT_MODE.replace(['reg oper account','org spec account','reg oper spec account','not specified'],[1,2,3,4])
application.HOUSETYPE_MODE=application.HOUSETYPE_MODE.replace(['block of flats','terraced house','specific housing'],[1,2,3])
application.WALLSMATERIAL_MODE=application.WALLSMATERIAL_MODE.replace(['Stone, brick','Block','Panel','Mixed','Wooden','Others','Monolithic'],[1,2,3,4,5,6,7])
application.EMERGENCYSTATE_MODE=application.EMERGENCYSTATE_MODE.replace(['Yes','No'],[1,2])


application.head()



# On vérifie maintenant qu'on a recodé toute les variables string en int que la corrélation peut se faire pour nos 106 colonnes
# On indique vouloir garder toutes les colonnes
pd.set_option('display.max_columns',None)
# On calcule la corrélation par paire des colonnes, à l'exclusion des valeurs NA/null.
coorMatrix = application.corr(method='pearson')
coorMatrix.shape

# Nous avaons bien 122 colonnes donc aucune perte d'information, nous allons maintenant étudier les corrélations sur 
# notre variable cible : TARGET -> 0 : client en difficulté de paiement ; 1 : autres cas 

# Calcul litérale 
# Le .abs nous permet de tout mettre en valeur absolue : si il y a un -1 (calcul négatif = anticorrélation) 
# ça donne aussi de l'information sur TARGET 

# On cherche a calculer les corrélations entre la variable cible : TARGET 
# et les autres variables en prenant compte des valeurs absolues
# On regarde un peu comment sont réparties les valeurs 
coorMatrix['TARGET'].abs().describe()


# On classe les valeurs par ordre descendant et on arrondit de 4 chiffres après la virgule. 
coorMatrix['TARGET'].sort_values (ascending = False)

# On importe nos corrélation dans un fichier csv
import csv
coorMatrix['TARGET'].to_csv('correlation_target_liste_desc.csv')

# On peut accéder à notre nouveau fichier 
correlation_target_liste_desc = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_MySQL/correlation_target_liste_desc.csv')#[['SK_ID_CURR','TARGET','NAME_CONTRACT_TYPE','CODE_GENDER','FLAG_OWN_CAR','FLAG_OWN_REALTY','CNT_CHILDREN','AMT_INCOME_TOTAL','AMT_CREDIT','AMT_ANNUITY']]#,'AMT_GOODS_PRICE','NAME_TYPE_SUITE','NAME_INCOME_TYPE','NAME_EDUCATION_TYPE','NAME_FAMILY_STATUS','NAME_HOUSING_TYPE','REGION_POPULATION_RELATIVE','DAYS_BIRTH','DAYS_EMPLOYED','DAYS_REGISTRATION','DAYS_ID_PUBLISH','OWN_CAR_AGE','FLAG_MOBIL','FLAG_EMP_PHONE','FLAG_WORK_PHONE','FLAG_CONT_MOBILE','FLAG_PHONE','FLAG_EMAIL','OCCUPATION_TYPE','CNT_FAM_MEMBERS','REGION_RATING_CLIENT','REGION_RATING_CLIENT_W_CITY','WEEKDAY_APPR_PROCESS_START','HOUR_APPR_PROCESS_START','REG_REGION_NOT_LIVE_REGION','REG_REGION_NOT_WORK_REGION','LIVE_REGION_NOT_WORK_REGION','REG_CITY_NOT_LIVE_CITY','REG_CITY_NOT_WORK_CITY','LIVE_CITY_NOT_WORK_CITY','ORGANIZATION_TYPE','EXT_SOURCE_1','EXT_SOURCE_2','EXT_SOURCE_3']]#'APARTMENTS_AVG	BASEMENTAREA_AVG	YEARS_BEGINEXPLUATATION_AVG	YEARS_BUILD_AVG	COMMONAREA_AVG	ELEVATORS_AVG	ENTRANCES_AVG	FLOORSMAX_AVG	FLOORSMIN_AVG	LANDAREA_AVG	LIVINGAPARTMENTS_AVG	LIVINGAREA_AVG	NONLIVINGAPARTMENTS_AVG	NONLIVINGAREA_AVG	APARTMENTS_MODE	BASEMENTAREA_MODE	YEARS_BEGINEXPLUATATION_MODE	YEARS_BUILD_MODE	COMMONAREA_MODE	ELEVATORS_MODE	ENTRANCES_MODE	FLOORSMAX_MODE	FLOORSMIN_MODE	LANDAREA_MODE	LIVINGAPARTMENTS_MODE	LIVINGAREA_MODE	NONLIVINGAPARTMENTS_MODE	NONLIVINGAREA_MODE	APARTMENTS_MEDI	BASEMENTAREA_MEDI	YEARS_BEGINEXPLUATATION_MEDI	YEARS_BUILD_MEDI	COMMONAREA_MEDI	ELEVATORS_MEDI	ENTRANCES_MEDI	FLOORSMAX_MEDI	FLOORSMIN_MEDI	LANDAREA_MEDI	LIVINGAPARTMENTS_MEDI	LIVINGAREA_MEDI	NONLIVINGAPARTMENTS_MEDI	NONLIVINGAREA_MEDI	FONDKAPREMONT_MODE	HOUSETYPE_MODE	TOTALAREA_MODE	WALLSMATERIAL_MODE	EMERGENCYSTATE_MODE	OBS_30_CNT_SOCIAL_CIRCLE	DEF_30_CNT_SOCIAL_CIRCLE	OBS_60_CNT_SOCIAL_CIRCLE	DEF_60_CNT_SOCIAL_CIRCLE	DAYS_LAST_PHONE_CHANGE	FLAG_DOCUMENT_2	FLAG_DOCUMENT_3	FLAG_DOCUMENT_4	FLAG_DOCUMENT_5	FLAG_DOCUMENT_6	FLAG_DOCUMENT_7	FLAG_DOCUMENT_8	FLAG_DOCUMENT_9	FLAG_DOCUMENT_10	FLAG_DOCUMENT_11	FLAG_DOCUMENT_12	FLAG_DOCUMENT_13	FLAG_DOCUMENT_14	FLAG_DOCUMENT_15	FLAG_DOCUMENT_16	FLAG_DOCUMENT_17	FLAG_DOCUMENT_18	FLAG_DOCUMENT_19	FLAG_DOCUMENT_20	FLAG_DOCUMENT_21	AMT_REQ_CREDIT_BUREAU_HOUR	AMT_REQ_CREDIT_BUREAU_DAY	AMT_REQ_CREDIT_BUREAU_WEEK	AMT_REQ_CREDIT_BUREAU_MON	AMT_REQ_CREDIT_BUREAU_QRT	AMT_REQ_CREDIT_BUREAU_YEAR]

               

application.info()

application.describe()

# On compte les valeurs manquantes en fonction des colonnes 
application.isnull().sum().head(60)

# On liste les colonnes qui on plus de 30% de valeurs manquantes 
cols_30_percent_missing_values = (application.columns[(round(100*(application.isnull().sum()/len(application.index)), 2)) >= 10]).tolist()
cols_30_percent_missing_values
