==> ipynb/6.1 POWER_BI_CLIENT.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt  
import seaborn as sns

# On importe le csv application_data
df = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_Originaux/application_data.csv')
# On indique vouloir garder toutes les colonnes
pd.set_option('display.max_columns',None)
df.head()

# On sélectionne les colonnes qui nous intéressent

df = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_Originaux/application_data.csv') [['SK_ID_CURR','TARGET','DAYS_BIRTH','CODE_GENDER','NAME_EDUCATION_TYPE','NAME_FAMILY_STATUS','CNT_CHILDREN','NAME_TYPE_SUITE','AMT_INCOME_TOTAL','NAME_INCOME_TYPE','OCCUPATION_TYPE','ORGANIZATION_TYPE','DAYS_EMPLOYED','FLAG_EMP_PHONE','FLAG_OWN_REALTY','NAME_HOUSING_TYPE','FLAG_OWN_CAR','DAYS_REGISTRATION','DAYS_ID_PUBLISH']]
df.rename(columns={'SK_ID_CURR': 'id_client', 'TARGET': 'cible','DAYS_BIRTH':'age','CODE_GENDER': 'genre',
                   'NAME_EDUCATION_TYPE':'niveau_etude','NAME_FAMILY_STATUS':'statut_matrimonial',
                   'CNT_CHILDREN':'nombre_enfants','NAME_TYPE_SUITE':'accompagnant',
                   'AMT_INCOME_TOTAL':'total_revenus','NAME_INCOME_TYPE':'type_revenu',
                   'OCCUPATION_TYPE':'categorie_pro','ORGANIZATION_TYPE':'type_entreprise', 
                   'DAYS_EMPLOYED':'anciennete','FLAG_EMP_PHONE':'telephone_pro',
                   'FLAG_OWN_REALTY':'proprietaire','NAME_HOUSING_TYPE':'type_logement',
                   'FLAG_OWN_CAR':'possede_voiture','DAYS_REGISTRATION':'anciennete_banque',
                   'DAYS_ID_PUBLISH':'mise_a_jour_documents',},inplace=True)

pd.set_option('display.max_columns',None)
df.head()
             

# On peut supprimer les NAN dans accompagnant (que 0.42 %)
# Aussi, on recode 'Other_A' et 'Other_B' en une variable 'Other' = 4
df = df.dropna(subset=['accompagnant'])
# On supprime les XNA
df = df[~(df.genre=='XNA')] # ou df = df[df.genre!=2].head()
# Dans la colonne categorie_pro, on remplace les NaN par la valeur 99
# df['categorie_pro'] = df['categorie_pro'].fillna()
df['categorie_pro'] = df['categorie_pro'].replace(np.nan, 99)
# On change les valeurs négatives en valeurs positives
# On convertit aussi les jours en années 
# Enfin, on convertit les float en integer. 
df['age']=abs(df['age']/365).astype(int).round()
df['anciennete']=abs(df['anciennete']/365).astype(int).round()
df['anciennete_banque']=abs(df['anciennete_banque']/365).astype(int).round()
df['mise_a_jour_documents']=abs(df['mise_a_jour_documents']/365).astype(int).round()
df['total_revenus']=(df['total_revenus']).astype(int)
# On peut supprimer les lignes avec des NaN 
df = df.dropna(subset=['total_revenus'])
# On remplace les valeurs abérantes par ne NaN
df.loc[df['anciennete'] >800, 'anciennete'] = np.nan
# On peut supprimer les lignes avec des NaN dans anceienneté (17.94%)
df = df.dropna(subset=['anciennete'])

df.head()

# On crée des catégories d'âge 

slots = ['20-30','30-40','40-50','50-60','60-70']
bins = [20,30,40,50,60,70]
df['age']=pd.cut(df['age'],bins,labels=slots)

# On vérifie le défaut de paiement en fonction des revenus
# On crée des quartiles

df['total_revenus'] = pd.qcut(df['total_revenus'], 10, labels = False)

# On recode les différents types d'entreprises pour avoir une liste moins grande
df.type_entreprise=df.type_entreprise.replace(['Business Entity Type 3','School','Government','Religion','Other','XNA','Electricity','Medicine','Business Entity Type 2','Self-employed','Transport: type 2','Construction','Housing','Kindergarten','Trade: type 7','Industry: type 11','Military','Services','Security Ministries','Transport: type 4','Industry: type 1','Emergency','Security','Trade: type 2','University','Transport: type 3','Police','Business Entity Type 1','Postal','Industry: type 4','Agriculture','Restaurant','Culture','Hotel','Industry: type 7','Trade: type 3','Industry: type 3','Bank','Industry: type 9','Insurance','Trade: type 6','Industry: type 2','Transport: type 1','Industry: type 12','Mobile','Trade: type 1','Industry: type 5','Industry: type 10','Legal Services','Advertising','Trade: type 5','Cleaning','Industry: type 13','Trade: type 4','Telecom','Industry: type 8','Realtor','Industry: type 6'],['Affaires','Public','Public','Religion','Autre','Autre','Primaire','Banque/Assurance','Affaires','Auto_Entrepreneur','Transport','Primaire','Entretien','Entretien','Commerce','Industrie','Public','Privé','Public','Transport','Industrie','Medecine','Public','Affaires','Public','Transport','Public','Affaires','Public','Industrie','Primaire','Privé','Public','Privé','Industrie','Commerce','Industrie','Banque/Assurance','Industrie','Banque/Assurance','Commerce','Industrie','Transport','Industrie','Privé','Commerce','Industrie','Industrie','Public','Privé','Commerce','Entretien','Industrie','Commerce','Privé','Industrie','Privé','Industrie'])


# On crée des catégories d'ancienneté au sein de la même entreprise 

slots = ['0-1','1-2','2-5','5-10','10-15','15-20','20 et +']
bins = [0,1,2,5,10,15,20,60]
df['anciennete']=pd.cut(df['anciennete'],bins,labels=slots)


# On crée des catégories d'ancienneté au sein de la même banque

slots = ['0-5','5-10','10-15','15-20','20-30','30 et +']
bins = [0,5,10,15,20,30,80]
df['anciennete_banque']=pd.cut(df['anciennete_banque'],bins,labels=slots)

df.head()

import csv
df.to_csv('POWER_BI_CLIENT.csv', index=False, quoting=csv.QUOTE_NONNUMERIC)
