==> ipynb/6.2 POWER_BI_CREDIT.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt  
import seaborn as sns

# On importe le csv previous_application
df = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_Originaux/previous_application.csv')
df.head()

# On séléctionne toutes les colonnes qui nous intéressent

df = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_Originaux/previous_application.csv')[['SK_ID_PREV','SK_ID_CURR','NAME_CONTRACT_TYPE','AMT_APPLICATION','AMT_GOODS_PRICE','NAME_CONTRACT_STATUS','DAYS_DECISION','NAME_PAYMENT_TYPE','CODE_REJECT_REASON','NAME_GOODS_CATEGORY']]


df.rename(columns={'SK_ID_PREV': 'id_credit',  
                    'SK_ID_CURR':'id_client',
                    'NAME_CONTRACT_TYPE':'type_credit',
                    'AMT_APPLICATION':'montant_demande_credit',
                    'AMT_GOODS_PRICE':'montant_credit_accorde',
                    'NAME_CONTRACT_STATUS':'statut_contrat',
                    'DAYS_DECISION':'nombre_jours_decision',
                    'NAME_PAYMENT_TYPE':'moyen_de_paiement',
                    'CODE_REJECT_REASON':'raison_rejet',
                    'NAME_GOODS_CATEGORY': 'type_bien'}, inplace=True)


pd.set_option('display.max_columns',None)
df.head()

# On peut supprimer les NaN dans montant_credit_accorde. On a un df initial avec 1,6 million d'entrées
df = df.dropna(subset=['montant_credit_accorde'])

# On recode les types de biens
df.type_bien=df.type_bien.replace(['XNA','Mobile','Consumer Electronics','Computers','Audio/Video','Furniture','Photo / Cinema Equipment','Construction Materials','Clothing and Accessories','Auto Accessories','Jewelry','Homewares','Medical Supplies','Vehicles','Sport and Leisure','Gardening','Other','Office Appliances','Tourism','Medicine','Direct Sales','Fitness','Additional Service','Education','Weapon','Insurance','Animals','House Construction'],['XNA','Téléphonie, Audio/Vidéo','Electronique','Téléphonie, Audio/Vidéo','Electronique','Maison','Electronique','Maison','Meubles','Articles ménagers','Vêtements et accessoirs','Maison','Bijoux','Articles ménagers','Médical','Maison','Loisirs','Maison','Loisirs','Bijoux','Loisirs','Médical','Loisirs','Loisirs','Loisirs','Loisirs','Loisirs','Maison'])


# On prend la valeur absolue des éléments de la colonne nombre_jours_décision
# On convertit aussi les jours en mois 
# Enfin, on convertit les nombres décimaux en entiers. 

df['nombre_jours_decision']=abs(df['nombre_jours_decision']/30)
df['montant_demande_credit']=df['montant_demande_credit'].astype(int)
df['montant_credit_accorde']=df['montant_credit_accorde'].astype(int)
# On renomme la colonne transformée
df.rename(columns={'nombre_jours_decision': 'nombre_mois_decision'}, inplace=True)
df['nombre_mois_decision'] = df['nombre_mois_decision'].astype(int)

df.head()

import csv
df.to_csv('POWER_BI_CREDIT.csv', index=False, quoting=csv.QUOTE_NONNUMERIC)
