==> ipynb/6.3 POWER_BI_MERGE.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt  
import seaborn as sns

# On importe le csv client POWER BI
dfclient = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_MySQL/POWER_BI_CLIENT.csv')
# On indique vouloir garder toutes les colonnes
pd.set_option('display.max_columns',None)
dfclient.head()

# On importe le csv credit
dfcredit = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_MySQL/POWER_BI_CREDIT.csv')
# On indique vouloir garder toutes les colonnes
pd.set_option('display.max_columns',None)
dfcredit.head()

dfmerge = pd.merge(dfclient, dfcredit, on="id_client")
dfmerge.head()

# On importe le csv type_credit
dftype_credit = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_MySQL/POWER_BI_TYPE_CREDIT.csv')

dftype_credit.head()

dfmerge1 = pd.merge(dfmerge,dftype_credit,  on="type_credit")
dfmerge1.head()

# On importe le csv bien
dfbien = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_MySQL/POWER_BI_TYPE_BIEN.csv')

dfbien.head()

dfmerge2 = pd.merge(dfmerge1,dfbien,  on="type_bien")
dfmerge2.head()

import csv
dfmerge2.to_csv('POWER_BI_MERGE_MySQL.csv', index=False, quoting=csv.QUOTE_NONNUMERIC)
