==> ipynb/7. Regression linéaire.ipynb <==
from sklearn.linear_model import LinearRegression
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt  
import seaborn as sns

# On importe le csv client
rl = pd.read_csv('/Users/p-asalisbury/Dropbox/Mon Mac (MacBook Air de P-A)/Desktop/Chef_doeuvre/CSV_MySQL/CSV_PREDICTION_MySQL.csv')
# On indique vouloir garder toutes les colonnes
pd.set_option('display.max_columns',None)
rl.head()

rl['statut_matrimonial'].value_counts()

rl = rl [['cible','niveau_etude','genre','age','statut_matrimonial','categorie_pro']]

from sklearn.neighbors import KNeighborsClassifier

model = KNeighborsClassifier() # On vient de charger un model de classification

y = rl['cible'] # On sélectionne la colonne cible
X = rl.drop('cible', axis=1) # On sélectionne toutes les colonnes sauf cible

model.fit(X,y)
model.score(X,y)

# Notre modèle fait une bonne prédiction dans 90 % des cas ! C'est-à-dire il arrive à dire si un client à eu un 
# défaut de paiement ou non. 

# On écrit une fonction pour prédir si notre client va avoir un défaut de paiement ou non.
# Secondary / secondary special = 0 Higher education = 1 Incomplete higher = 2 Lower secondary = 3 Academic degree = 4
# Laborers = 0 Core staff = 1 Accountants = 2 Managers = 3 Drivers = 4 Sales staff = 5 Cleaning staff = 6 Cooking 
# 7 Private service staff = 8 Medicine staff = 9 Security staff = 10 High skill tech staff = 11Waiters/barmen staff 
# 12 Low-skill Laborers = 13 Realty agents = 14 Secretaries = 15 IT staff = 16 HR staf = 17

def defaut(model,niveau_etude=2, genre=0, age=31, statut_matrimonial=0, categorie_pro=0  ):
    x = np.array([niveau_etude, genre, age, statut_matrimonial, categorie_pro]).reshape(1,5)
    print(model.predict(x))
    print(model.predict_proba(x))

defaut(model) # On estime si je vais avoir un défaut de paiement mais ce n'est pas suffisant ..
              # .. on calcule la probabilité que j'ai un défaut de paiement

# On a 80 % de chance d'honnorer son prêt et 20 % d'avoir un défaut de paiement. 
# Pour un homme de 31 ans célibataire et ouvrier.
